chore: restructure benches.
