+++ src/tests/mod.rs
@@ -37,9 +37,10 @@ pub mod test_helpers {
             .into_iter()
             .map(|x| EF::new(x.get_val(), Some(num_levels)))
             .collect::<Vec<EF>>();
-        let eq_challenges = if algorithm == AlgorithmType::PrecomputationWithEq
-            || algorithm == AlgorithmType::ToomCookWithEq
-        {
+
+        let is_algo_with_eq_poly = algorithm == AlgorithmType::PrecomputationWithEq
+            || algorithm == AlgorithmType::ToomCookWithEq;
+        let eq_challenges = if is_algo_with_eq_poly {
             let eq_challenges = EF::rand_vector(nv, Some(num_levels));
             let eq_challenges_evals = EqPoly::new(eq_challenges.clone()).compute_evals(false);
             polynomial_hadamard_ef
@@ -55,13 +56,7 @@ pub mod test_helpers {
             .iter()
             .fold(EF::zero(), |acc, ph| acc + ph.clone());
 
-        let sumcheck_degree = if algorithm == AlgorithmType::PrecomputationWithEq
-            || algorithm == AlgorithmType::ToomCookWithEq
-        {
-            degree + 1
-        } else {
-            degree
-        };
+        let sumcheck_degree = is_algo_with_eq_poly.then(|| degree + 1).unwrap_or(degree);
         let prover_state: ProverState<EF, BF> =
             IPForMLSumcheck::<EF, BF>::prover_init(&polynomials, sumcheck_degree, algorithm);
